+++ package.json
@@ -23,6 +23,7 @@
   "homepage": "https://github.com/kaathryyn/2022_SIS_Team2#readme",
   "dependencies": {
     "@google-cloud/vision": "^2.4.2",
-    "express": "^4.18.1"
+    "express": "^4.18.1",
+    "socket.io": "^4.5.1"
   }
 }
